Project_01_Vehicle_Insurance_Data/requirements.txt
ipykernel
pandas
numpy
matplotlib
plotly
seaborn
scikit-learn
pymongo
from_root
dill
certifi
PyYAML
boto3
mypy-boto3-s3
botocore
fastapi
python-multipart
uvicorn
jinja2
imblearn
python-dotenv
huggingface_hub
-e .